+++ src/components/modal/ModalCircleSettings.svelte
@@ -12,17 +12,14 @@
 
 	let _settings: CircleSettings;
 	let _onOK: (settings: CircleSettings) => void;
-	let _onDelete: (() => void) | undefined;
 	let isOpen: boolean = false;
 
 	export const open = (
 		settings: CircleSettings,
-		onOK: (settings: CircleSettings) => void,
-		onDelete?: () => void
+		onOK: (settings: CircleSettings) => void
 	): void => {
 		_settings = settings;
 		_onOK = onOK;
-		_onDelete = onDelete;
 		isOpen = true;
 	};
 </script>
@@ -51,15 +48,5 @@
 			class="me-2">OK</Button
 		>
 		<Button on:click={() => (isOpen = false)} color="alternative" class="me-2">Cancel</Button>
-		{#if _onDelete}
-			<Button
-				on:click={() => {
-					isOpen = false;
-					_onDelete && _onDelete();
-				}}
-				color="red"
-				class="me-2 float-right">Remove</Button
-			>
-		{/if}
 	</div>
 </Modal>